+++ src/config.js
@@ -1,6 +1,6 @@
 var Config = {
     DEBUG: false,
-    LIB_VERSION: '2.9.5'
+    LIB_VERSION: '2.9.10'
 };
 
 export default Config;